Move report provider to aw project. #5349
