+++ modules/AlphaWalletFoundation/AlphaWalletFoundation/Tokens/ClientSideTokenSourceProvider.swift
@@ -111,7 +111,7 @@ public class ClientSideTokenSourceProvider: TokenSourceProvider {
 
 extension ClientSideTokenSourceProvider: TokenBalanceFetcherDelegate {
     public func didUpdateBalance(value actions: [AddOrUpdateTokenAction], in fetcher: TokenBalanceFetcher) {
-        crashlytics?.logLargeNftJsonFiles(for: actions, fileSizeThreshold: 10)
+        crashlytics.logLargeNftJsonFiles(for: actions, fileSizeThreshold: 10)
         tokensDataStore.addOrUpdate(actions)
     }
 }
